Adding the project
Visualization of the massive parallel data
The data is a simulation of the explosion in different timesteps, and
the software is capable of displaying it while loading in real time(the
data is approx. 56M points, what results in ~6 GB of text files, which
contain 7 different values for the given points in the mesh grid)

The software still needs some fixes to be more universal to plot the
data, for now it's working on a fixed amount of data of 384x384x384
points.